refactor: Line 도메인에서 Sections 분리
