+++ src/test/java/nextstep/subway/line/application/LineServiceTest.java
@@ -39,6 +39,6 @@ class LineServiceTest {
 
         // then
         Line line = lineService.findLineById(이호선.getId());
-        assertThat(line.getSections()).hasSize(2);
+        assertThat(line.getSections().size()).isEqualTo(2);
     }
 }
